requirements.txt
requests
websocket-client
python-dotenv
sentence-transformers
torch
py-clob-client
